refactor(providers): settings types and defaults live in their provider modules
Move each provider's Settings interface and defaultXxxSettings const out
of settingsStore.ts into that provider's own config module, and delete
the dead ProviderConfig.defaults field (zero consumers) along with the
matching defaults block in every getConfig(). settingsStore.ts now
imports the moved declarations and re-exports the types so existing
importers (MainPanel, LanguageSection, etc.) keep compiling unchanged.
Pure declaration motion — no behavior or value changes.
